Optimized GameRecap by breaking out recap text #40
