+++ src/components/GameRecap/GameRecap.tsx
@@ -3,54 +3,19 @@ import { Choice } from '../../types/Image.types';
 import PhotoCarousel from '../PhotoCarousel/PhotoCarousel';
 import ShareButton from '../ShareButton/ShareButton';
 import { useEffect, useMemo, useState } from 'react';
-import { currentDay } from '../../services/Day.service';
+import { getToday } from '../../services/Day.service';
 import { generateScoreText } from '../../services/Score.service';
 import { useSpring, animated } from '@react-spring/web';
-import ConfettiExplosion from 'react-confetti-explosion';
+import { recapAnimationTime } from '../../constants/GameRecapConstants';
+import GameRecapText from '../GameRecapText.tsx/GameRecapText';
+import { getLastChoiceKeeper } from '../../services/Choices.service';
 
-const day = currentDay();
-const animationTime = 500;
+const day = getToday();
 
 const GameRecap = () => {
   const choices = useChoiceKeeper();
-  const score = useMemo(() => {
-    return choices.filter((choice) => choice.isCorrect).length;
-  }, [choices]);
   const scoreText = useMemo(() => generateScoreText(choices, day), [choices]);
-  const [showConfetti, setShowConfetti] = useState(false);
-
-  const translateAnimation = useSpring({
-    from: {
-      transform: `translateY(20px)`,
-    },
-    to: { transform: `translateY(0px)` },
-    config: {
-      duration: animationTime,
-      easing: (t) => Math.pow(t - 1, 3) + 1,
-    },
-  });
-
-  const opacityAnimation = useSpring({
-    from: {
-      opacity: 0,
-    },
-    to: { opacity: 1 },
-    config: {
-      duration: animationTime,
-      easing: (t) => Math.pow(t, 3),
-    },
-  });
-
-  const animations = {
-    ...translateAnimation,
-    ...opacityAnimation,
-  };
-
-  useEffect(() => {
-    setTimeout(() => {
-      setShowConfetti(score === choices.length);
-    }, animationTime);
-  }, [score, choices]);
+  const animations = useRecapAnimations();
 
   return (
     <animated.div className='w-full h-full' style={animations}>
@@ -61,21 +26,7 @@ const GameRecap = () => {
         className='text-center'
         vertical
       >
-        <div className='relative'>
-          {showConfetti && (
-            <div className='absolute top-0 left-1/2 transform -translate-x-1/2 z-50'>
-              <ConfettiExplosion
-                force={0.6}
-                duration={3000}
-                particleCount={160}
-                width={1000}
-              />
-            </div>
-          )}
-          <p className='text-2xl'>
-            You got {score} out of {choices.length} correct!
-          </p>
-        </div>
+        <GameRecapText choices={choices} />
         <div className='w-11/12'>
           <PhotoCarousel choices={choices} />
         </div>
@@ -89,13 +40,42 @@ const useChoiceKeeper = () => {
   const [choiceKeeper, setChoiceKeeper] = useState<Choice[]>([]);
 
   useEffect(() => {
-    const savedChoices = localStorage.getItem('last_choice_keeper');
+    const savedChoices = getLastChoiceKeeper();
     if (savedChoices) {
-      setChoiceKeeper(JSON.parse(savedChoices));
+      setChoiceKeeper(savedChoices);
     }
   }, []);
 
   return choiceKeeper;
+};
+
+const useRecapAnimations = () => {
+  const translateAnimation = useSpring({
+    from: {
+      transform: `translateY(20px)`,
+    },
+    to: { transform: `translateY(0px)` },
+    config: {
+      duration: recapAnimationTime,
+      easing: (t) => Math.pow(t - 1, 3) + 1,
+    },
+  });
+
+  const opacityAnimation = useSpring({
+    from: {
+      opacity: 0,
+    },
+    to: { opacity: 1 },
+    config: {
+      duration: recapAnimationTime,
+      easing: (t) => Math.pow(t, 3),
+    },
+  });
+
+  return {
+    ...translateAnimation,
+    ...opacityAnimation,
+  };
 };
 
 export default GameRecap;
