+++ src/constants/GameRecapConstants.ts
@@ -1,0 +1,1 @@
+export const recapAnimationTime = 500;
